fix: 할 일 목록과 타이머 데이터 동시 동기화하도록 변경
