+++ src/components/study-detail/main/TodoListSection.tsx
@@ -1,24 +1,31 @@
-import useTodoList from '../../../hooks/study-detail/todo/useTodoList';
 import PlayIcon from '../../../assets/icons/play.svg';
 import CloseIcon from '../../../assets/icons/close.svg';
 import PauseIcon from '../../../assets/icons/pause.svg';
 import CreateTodo from './CreateTodo';
 import useTodoActions from '../../../hooks/study-detail/todo/useTodoActions';
-import { TodoType, TimerType } from '../../../types/interface';
+import { TodoListType, TodoType } from '../../../types/interface';
 import { formatTime } from '../../../utils/commonUtils';
 import useTodoTimer from '../../../hooks/study-detail/todo/useTodoTimer';
 import { Dispatch, SetStateAction } from 'react';
+import { TimerType } from '../../../types/interface';
 
 export default function TodoListSection({
   studyId,
   userId,
+  todoList,
   setTimers,
+  todoStates,
+  handleCheckboxChange,
+  setLocalTodoList,
 }: {
   studyId: number;
   userId: number;
+  todoList: TodoListType;
   setTimers: Dispatch<SetStateAction<TimerType[]>>;
+  todoStates: Record<number, boolean>;
+  handleCheckboxChange: (todoId: number, isCompleted: boolean) => void;
+  setLocalTodoList: Dispatch<SetStateAction<TodoListType>>;
 }) {
-  const { todoList, todoListLoading, todoStates, handleCheckboxChange, setLocalTodoList } = useTodoList(studyId);
   const {
     isAdding,
     setIsAdding,
@@ -32,7 +39,6 @@ export default function TodoListSection({
     setEditingTodo,
   } = useTodoActions(studyId);
   const { startTimer, stopTimer, activeTodoId } = useTodoTimer({ studyId, userId, setTimers, setLocalTodoList });
-  if (todoListLoading || !todoList || !todoList.todos) return null;
   const { studyTotalTime, todos } = todoList;
 
   return (
